+++ src/nft/utils/asset.ts
@@ -1,0 +1,18 @@
+import { GenieAsset } from 'nft/types'
+
+export function getRarityStatus(
+  rarityStatusCache: Map<string, boolean>,
+  id: string,
+  assets?: (GenieAsset | undefined)[]
+) {
+  if (rarityStatusCache.has(id)) {
+    return rarityStatusCache.get(id)
+  }
+  const hasRarity = assets && Array.from(assets).reduce((reducer, asset) => !!(reducer || asset?.rarity), false)
+
+  if (hasRarity) {
+    rarityStatusCache.set(id, hasRarity)
+  }
+
+  return hasRarity
+}
